requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
boto3==1.34.25
botocore==1.34.25
Django==4.0.1
django-debug-toolbar==4.2.0
django-environ==0.8.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.62
redis==5.0.1
requests==2.31.0